+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Changes
 
 * Rename the `symbol` param in the `undef` middleware to `sym`. (the old name is retained for backwards compatibility)
+* Rename the `symbol` param in the `complete` middleware to `prefix`. (the old name is retained for backwards compatibility)
 
 ## 0.24.0 (2020-02-14)